refactor: 컴포넌트 내보내기를 한 곳에 병합
