+++ src/components/index.ts
@@ -1,0 +1,8 @@
+import Button from './Button';
+import CategoryImage from './CategoryImage';
+import Header from './Header';
+import RestaurantDetail from './RestaurantDetail';
+import RestaurantItem from './RestaurantItem';
+import RestaurantList from './RestaurantList';
+
+export { Button, CategoryImage, Header, RestaurantDetail, RestaurantList, RestaurantItem };
